+++ source/BaseEngine/Terrain/Terrain.cpp
@@ -137,11 +137,11 @@ void CTerrain::CreateTerrain()
 
 	int pointer = 0;
 	//Triaagnle strip
-	for (int gz = 0; gz <m_VertexCount - 1; gz++)
+	for (unsigned short gz = 0; gz < m_VertexCount - 1; gz++)
 	{
 		if ((gz & 1) == 0) 
 		{ // even rows
-			for (int gx = 0; gx <m_VertexCount - 1; gx++)
+			for (unsigned short  gx = 0; gx < m_VertexCount - 1; gx++)
 			{
 				m_Indices.push_back(gx + gz * m_VertexCount);
 				m_Indices.push_back(gx + (gz + 1) * m_VertexCount);
@@ -149,7 +149,7 @@ void CTerrain::CreateTerrain()
 		}
 		else
 		{ // odd rows
-			for (int gx = m_VertexCount - 1; gx>0; gx--)
+			for (unsigned short  gx = m_VertexCount - 1; gx>0; gx--)
 			{
 				m_Indices.push_back(gx + (gz + 1) * m_VertexCount);
 				m_Indices.push_back(gx - 1 + +gz * m_VertexCount);
